fix: ./config and ./plugins confusion
This is so confusing, ./config should only have the defined files,
lazyvim, options, automap and keymaps.

The ./plugins is where you can put any .lua file you want and it is
executed.

The confusion is that with ./plugins, the format is special, you just
return a dictionary and it does the require("_your plugin_").setup() for
you, but in ./config, if you put it into lazy.config then you need to do
this yourself